Editorial changes - terminology issue ('gesture')
Terminology: It is quite common to see simple touch operations like a single tap referred to as gesture. For me it is not clear where 'gesture' starts. I don't think it is evident and common-sense that 'gesture' would not include single tap activation. So talking about a "non-gesture-based alternative" seems problematic if we mwean that theae should be controls that can be activated by single tap (or, with AT on, swipe followed double tap or alternatively, a split tap).

Consequently I have tried to reword verious parts of this SC to refer to "complex or timed pointer gestures or multi-pointer gestures" as opposed to simple pointer gesture (such as tapping or a mouse click).

Do we need to spell out that this SC applies to touch interfaces with built-in AT not active? Or have it apply generally but spell out that complex gestures usuall do not work when AT is on unless pass-through gestures are used (which often do not work reliably)?

Another issue:  Note 2 refers to "web content" - does this preclude applicability of the SC to native apps?